Add HistoryItem, move popover
