--- recipe-server/client/control/tests/components/recipes/test_RevisionApprovalTag.js
@@ -1,20 +1,0 @@
-import { Map } from 'immutable';
-import React from 'react';
-import { shallow } from 'enzyme';
-
-import TestComponent from 'control/components/recipes/RevisionApprovalTag';
-
-const { WrappedComponent: RevisionApprovalTag } = TestComponent;
-
-describe('<RevisionApprovalTag>', () => {
-  const props = {
-    revision: new Map(),
-    status: null,
-  };
-
-  it('should work', () => {
-    const wrapper = () => shallow(<RevisionApprovalTag {...props} />);
-
-    expect(wrapper).not.toThrow();
-  });
-});
