+++ recipe-server/client/control_new/less/components/history-timeline.less
@@ -26,6 +26,12 @@
   width: 8px;
 }
 
+.ant-timeline-item:last-child {
+  .ant-timeline-item-tail {
+    height: 5px;
+  }
+}
+
 .ant-timeline-item-head-custom {
   top: 1px;
 }
